# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: WhatsApp Agent Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /widgets:
    get:
      operationId: getWidgets
      tags: [widgets]
      summary: List all widgets
      responses:
        "200":
          description: List of widgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Widget"
    post:
      operationId: createWidget
      tags: [widgets]
      summary: Create a widget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WidgetInput"
      responses:
        "201":
          description: Created widget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Widget"
  /widgets/{id}:
    patch:
      operationId: updateWidget
      tags: [widgets]
      summary: Update a widget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WidgetUpdate"
      responses:
        "200":
          description: Updated widget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Widget"
    delete:
      operationId: deleteWidget
      tags: [widgets]
      summary: Delete a widget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Profile
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # Organization
  /organization:
    get:
      operationId: getOrganization
      tags: [organization]
      summary: Get organization info
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    patch:
      operationId: updateOrganization
      tags: [organization]
      summary: Update organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdate"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organization/members:
    get:
      operationId: getOrganizationMembers
      tags: [organization]
      summary: Get organization members
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrgMember"
    post:
      operationId: inviteMember
      tags: [organization]
      summary: Invite a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInviteInput"
      responses:
        "201":
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgMember"

  # API Keys
  /api-keys:
    get:
      operationId: getApiKeys
      tags: [apikeys]
      summary: List API keys
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: createApiKey
      tags: [apikeys]
      summary: Generate a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "201":
          description: Created API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
  /api-keys/{id}:
    delete:
      operationId: deleteApiKey
      tags: [apikeys]
      summary: Delete an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Widget:
      type: object
      required: [id, name, phoneNumber, welcomeText, buttonColor, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phoneNumber:
          type: string
        welcomeText:
          type: string
        buttonColor:
          type: string
        buttonText:
          type: ["string", "null"]
        position:
          type: string
          enum: [bottom-right, bottom-left]
        isActive:
          type: boolean
        embedCode:
          type: ["string", "null"]
        createdAt:
          type: string
    WidgetInput:
      type: object
      required: [name, phoneNumber, welcomeText, buttonColor]
      properties:
        name:
          type: string
        phoneNumber:
          type: string
        welcomeText:
          type: string
        buttonColor:
          type: string
        buttonText:
          type: string
        position:
          type: string
          enum: [bottom-right, bottom-left]
    WidgetUpdate:
      type: object
      properties:
        name:
          type: string
        phoneNumber:
          type: string
        welcomeText:
          type: string
        buttonColor:
          type: string
        buttonText:
          type: string
        position:
          type: string
          enum: [bottom-right, bottom-left]
        isActive:
          type: boolean
    UserProfile:
      type: object
      required: [id, name, email, plan, creditsUsed, creditsTotal, activeAgents, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        plan:
          type: string
        creditsUsed:
          type: integer
        creditsTotal:
          type: integer
        activeAgents:
          type: integer
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
    Organization:
      type: object
      required: [id, name, plan, memberCount, customDomain]
      properties:
        id:
          type: integer
        name:
          type: string
        plan:
          type: string
        memberCount:
          type: integer
        customDomain:
          type: ["string", "null"]
    OrganizationUpdate:
      type: object
      properties:
        name:
          type: string
        customDomain:
          type: string
    OrgMember:
      type: object
      required: [id, name, email, role, joinedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, admin, member]
        joinedAt:
          type: string
    MemberInviteInput:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
        role:
          type: string
          enum: [admin, member]
    ApiKey:
      type: object
      required: [id, name, keyPreview, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        keyPreview:
          type: string
        fullKey:
          type: ["string", "null"]
        createdAt:
          type: string
    ApiKeyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string